+++ demos/chinook/src/main/java/is/codion/demos/chinook/manual/FrameworkModelDemo.java
@@ -207,7 +207,7 @@ public final class FrameworkModelDemo {
 
 			return connection.select(and(
 							Customer.EMAIL.isNotNull(),
-							queryModel.condition().where(Conjunction.AND))
+							queryModel.condition().where())
 			);
 		});
 		// end::customDataSource[]